fix: Isolated SummaryActionItems, SummaryKeyPoints, getStatusColor, SummaryMetadata into summaryInfoComponents.tsx from app/workspace/[id]/meeting-summaries/page.tsx to prevent potential error.
