+++ src/app/page.tsx
@@ -1,3 +1,11 @@
+"use client"
+import {useEffect} from "react";
+import {useRouter} from "next/navigation";
+
 export default function Home() {
-  return <h1 className="text-3xl font-bold underline">Hello world!</h1>;
+  const router = useRouter();
+  useEffect(() => {
+    router.push("/landing");
+  })
+  return <></>
 }
